refactor(View): 重新導向首頁 - 首頁來不及製作，改用PLP
refactor(AdminCenter): UI小修

feat(AdminCenter): 新增內容
- 基本書本管理
- 標籤管理
- 分類管理

+++ SpecialTopic/UsedBooks/Backend/DTOs/BookCardDto.cs
@@ -4,7 +4,7 @@
     {
         public int BookID { get; set; }
         public string BookName { get; set; } = string.Empty;
-        public decimal SalePrice { get; set; }
+        public int SalePrice { get; set; }
         public string Authors { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
         public string ImagePath { get; set; } = string.Empty;
